pyproject.toml
[project]
name = "assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "daphne>=4.2.1",
    "django>=5.2.11",
    "django-elasticsearch-dsl==8.0.0",
    "django-environ>=0.13.0",
    "django-tenant-schemas>=2.0.0",
    "djangorestframework>=3.16.1",
    "elasticsearch==8.12.0",
    "elasticsearch-dsl==8.12.0",
    "faker>=40.5.1",
    "psycopg-binary>=3.3.3",
    "psycopg2-binary>=2.9.11",
    "pytest-django>=4.12.0",
    "pytest-mock>=3.15.1",
    "redis>=7.2.0",
    "tqdm>=4.67.3",
    "django-cors-headers>=3.13.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "isort>=8.0.1",
]
